build: update go directive to released version
Per https://github.com/dependabot/dependabot-core/issues/11825 and
https://github.com/dependabot/dependabot-core/pull/11826 ,
the dependabot has been updated to use GOTOOLCHAIN=go1.24.1 and it will add toolchain
in go mod tidy when the go directive is specified with language version(1.N). As we don't
want to the toolchain to be specified with go1.24.1, change the go directive to
released version(1.N.P).

Signed-off-by: Jude Hung <jude@iotechsys.com>